+++ src/lib/pages/blocks/index.tsx
@@ -5,6 +5,7 @@ import { AmpEvent, track } from "lib/amplitude";
 import { useTierConfig } from "lib/app-provider";
 import PageContainer from "lib/components/PageContainer";
 import { PageHeader } from "lib/components/PageHeader";
+import { CelatoneSeo } from "lib/components/Seo";
 
 import { RecentBlocksTable } from "./components/RecentBlocksTable";
 
@@ -19,6 +20,7 @@ const BlocksPage = () => {
 
   return (
     <PageContainer>
+      <CelatoneSeo pageName="Blocks" />
       <PageHeader
         title="Blocks"
         subtitle="This page displays all blocks on this network sorted by recency"