+++ src/lib/pages/interact/index.tsx
@@ -1,14 +1,20 @@
-import { Box, Button, Flex, Grid, Text, useDisclosure } from "@chakra-ui/react";
+import { Box, Button, Flex, Text, useDisclosure } from "@chakra-ui/react";
 import { useRouter } from "next/router";
+import type { Dispatch, SetStateAction } from "react";
 import { useCallback, useEffect, useState } from "react";
 
 import { AmpEvent, track, trackToModuleInteraction } from "lib/amplitude";
-import { useInternalNavigate, useMoveConfig } from "lib/app-provider";
+import {
+  useInternalNavigate,
+  useMobile,
+  useMoveConfig,
+} from "lib/app-provider";
 import { CustomIcon } from "lib/components/icon";
 import { LabelText } from "lib/components/LabelText";
-import { ModuleSourceCode } from "lib/components/module";
+import { FunctionCard, ModuleSourceCode } from "lib/components/module";
 import PageContainer from "lib/components/PageContainer";
 import { PageHeader } from "lib/components/PageHeader";
+import { CelatoneSeo } from "lib/components/Seo";
 import { useOpenNewTab } from "lib/hooks";
 import {
   useModuleByAddressLcd,
@@ -18,26 +24,122 @@ import type { Addr, ExposedFunction, IndexedModule } from "lib/types";
 import { getFirstQueryParam } from "lib/utils";
 
 import {
-  FunctionSelectBody,
-  FunctionSelectPanel,
+  InteractionBodySection,
+  InteractionBodySectionMobile,
   InteractionTabs,
   ModuleSelectDrawer,
   ModuleSelectDrawerTrigger,
 } from "./component";
+import { ModuleSelectDrawerMobile } from "./component/drawer/ModuleSelectDrawerMobile";
+import { SelectedFunctionCard } from "./component/SelectedFunctionCard";
+import { ModuleInteractionMobileStep } from "./types";
+
+const FunctionSection = ({
+  selectedFn,
+  setSelectedFn,
+  isZeroState,
+  handleDrawerOpen,
+}: {
+  selectedFn?: ExposedFunction;
+  setSelectedFn?: Dispatch<SetStateAction<ExposedFunction | undefined>>;
+  isZeroState?: boolean;
+  handleDrawerOpen?: (step: ModuleInteractionMobileStep) => void;
+}) => (
+  <Flex
+    borderTop="1px solid"
+    borderTopColor="gray.700"
+    pt={4}
+    mt={2}
+    direction="column"
+    gap={4}
+  >
+    <Flex direction="column" gap={2}>
+      <Flex gap={2} alignItems="center">
+        <Text as="h6" variant="h6" fontWeight={600}>
+          Selected Function
+        </Text>
+        <Button
+          size="sm"
+          variant={isZeroState ? "ghost-gray" : "ghost-primary"}
+          leftIcon={<CustomIcon name="swap" boxSize={3} />}
+          px={1}
+          onClick={() =>
+            handleDrawerOpen?.(ModuleInteractionMobileStep.SelectFunction)
+          }
+        >
+          Change Function
+        </Button>
+      </Flex>
+      {selectedFn && setSelectedFn && (
+        <FunctionCard
+          variant="readonly"
+          exposedFn={selectedFn}
+          isReadOnly
+          onFunctionSelect={() => {
+            setSelectedFn(selectedFn);
+          }}
+        />
+      )}
+    </Flex>
+    {selectedFn ? (
+      <SelectedFunctionCard fn={selectedFn} />
+    ) : (
+      <Flex
+        p={4}
+        borderRadius={8}
+        alignItems="center"
+        justifyContent="center"
+        bg="background.main"
+        color="text.dark"
+      >
+        Please select module first
+      </Flex>
+    )}
+  </Flex>
+);
+
+const ZeroState = ({
+  onOpen,
+  isMobile,
+}: {
+  onOpen: () => void;
+  isMobile: boolean;
+}) => (
+  <Flex
+    direction={{ base: "column", md: "row" }}
+    alignItems={{ md: "center" }}
+    justifyContent={{ md: "space-between" }}
+    w="full"
+    gap={4}
+  >
+    <p>Select a module to interact with ...</p>
+    <ModuleSelectDrawerTrigger triggerVariant="select-module" onOpen={onOpen} />
+    {isMobile && <FunctionSection isZeroState />}
+  </Flex>
+);
 
 export const Interact = () => {
   useMoveConfig({ shouldRedirect: true });
 
   const router = useRouter();
+  const isMobile = useMobile();
   const navigate = useInternalNavigate();
-  const { isOpen, onClose, onOpen } = useDisclosure();
-  const openNewTab = useOpenNewTab();
 
+  const openNewTab = useOpenNewTab();
+  const { isOpen, onClose, onOpen } = useDisclosure();
   const [module, setModule] = useState<IndexedModule>();
   const [selectedType, setSelectedType] = useState<InteractionTabs>(
     InteractionTabs.VIEW_MODULE
   );
   const [selectedFn, setSelectedFn] = useState<ExposedFunction>();
+
+  const [step, setStep] = useState(ModuleInteractionMobileStep.SelectModule);
+  const [selectedModule, setSelectedModule] = useState<IndexedModule>();
+
+  const handleDrawerOpen = (selectedStep: ModuleInteractionMobileStep) => {
+    setStep(selectedStep);
+    onOpen();
+  };
 
   const handleSetSelectedType = (type: string) =>
     setSelectedType(
@@ -47,16 +149,16 @@ export const Interact = () => {
     );
 
   const handleModuleSelect = useCallback(
-    (selectedModule: IndexedModule, fn?: ExposedFunction) => {
-      setModule(selectedModule);
+    (selectedModuleInput: IndexedModule, fn?: ExposedFunction) => {
+      setModule(selectedModuleInput);
       setSelectedFn(fn);
       handleSetSelectedType(fn?.is_view ?? true ? "view" : "execute");
 
       navigate({
         pathname: "/interact",
         query: {
-          address: selectedModule.address,
-          moduleName: selectedModule.moduleName,
+          address: selectedModuleInput.address,
+          moduleName: selectedModuleInput.moduleName,
           ...(fn && { functionName: fn.name }),
         },
         options: { shallow: true },
@@ -134,15 +236,27 @@ export const Interact = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [router.isReady]);
 
+  useEffect(() => {
+    if (!module || !isMobile || !router.isReady) return;
+
+    setSelectedModule(module);
+    handleDrawerOpen(ModuleInteractionMobileStep.SelectFunction);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [router.isReady, module]);
+
   return (
     <>
+      <CelatoneSeo pageName="View / Execute Module" />
       <PageContainer
         overflow="hidden"
         minH="unset"
         display="grid"
         gridTemplateRows="auto auto 1fr"
       >
-        <PageHeader title=" Module Interactions" docHref="move/view-execute" />
+        <PageHeader
+          title={isMobile ? "View Module" : "Module Interactions"}
+          docHref="move/view-execute"
+        />
         <Flex
           alignItems="center"
           justifyContent="space-between"
@@ -152,35 +266,62 @@ export const Interact = () => {
           mb={8}
         >
           {module ? (
-            <>
+            <Flex
+              direction={{ base: "column", md: "row" }}
+              alignItems={{ md: "center" }}
+              justifyContent={{ md: "space-between" }}
+              w="full"
+              gap={{ base: 4, md: 0 }}
+            >
               <Flex direction="column" gap={4}>
-                <LabelText label="Module Path" labelWeight={600}>
-                  <Flex align="center" gap={1}>
-                    <Text variant="body1">{module.address.toString()}</Text>
+                <LabelText
+                  label="Module Path"
+                  labelWeight={600}
+                  labelColor="text.main"
+                >
+                  <Flex align="center" gap={1} display="inline">
+                    <Text
+                      variant="body1"
+                      display="inline-flex"
+                      wordBreak="break-all"
+                    >
+                      {module.address}
+                    </Text>
                     <CustomIcon
                       name="chevron-right"
                       color="gray.600"
                       boxSize={3}
                     />
-                    <Text variant="body1" fontWeight={700}>
+                    <Text
+                      variant="body1"
+                      fontWeight={700}
+                      display="inline-flex"
+                    >
                       {module.moduleName}
                     </Text>
                   </Flex>
                 </LabelText>
-                <LabelText label="Friends" labelWeight={600}>
+                <LabelText
+                  label="Friends"
+                  labelWeight={600}
+                  labelColor="text.main"
+                >
                   <Text variant="body2" color="gray.400">
                     {JSON.stringify(module.parsedAbi.friends)}
                   </Text>
                 </LabelText>
               </Flex>
-              <Flex direction="column" gap={2}>
+              <Flex direction={{ base: "row", md: "column" }} gap={2}>
                 <ModuleSelectDrawerTrigger
                   triggerVariant="change-module"
                   buttonText="Change Module"
-                  onOpen={onOpen}
+                  onOpen={() =>
+                    handleDrawerOpen(ModuleInteractionMobileStep.SelectModule)
+                  }
                 />
                 <Button
                   variant="ghost-gray"
+                  size={{ base: "sm", md: "md" }}
                   rightIcon={
                     <CustomIcon name="launch" boxSize={3} color="text.dark" />
                   }
@@ -199,43 +340,53 @@ export const Interact = () => {
                   See Module
                 </Button>
               </Flex>
-            </>
+              {isMobile && selectedFn && (
+                <FunctionSection
+                  selectedFn={selectedFn}
+                  setSelectedFn={setSelectedFn}
+                  handleDrawerOpen={handleDrawerOpen}
+                />
+              )}
+            </Flex>
           ) : (
-            <>
-              <p>Select a module to interact with ...</p>
-              <ModuleSelectDrawerTrigger
-                triggerVariant="select-module"
-                onOpen={onOpen}
-              />
-            </>
+            <ZeroState onOpen={onOpen} isMobile={isMobile} />
           )}
-          <ModuleSelectDrawer
-            isOpen={isOpen}
-            onClose={onClose}
-            hexAddress={module?.address}
-            handleModuleSelect={handleModuleSelect}
-          />
+          {isMobile ? (
+            <ModuleSelectDrawerMobile
+              isOpen={isOpen}
+              onClose={onClose}
+              hexAddress={module?.address}
+              handleModuleSelect={handleModuleSelect}
+              step={step}
+              setStep={setStep}
+              selectedModule={selectedModule}
+              setSelectedModule={setSelectedModule}
+            />
+          ) : (
+            <ModuleSelectDrawer
+              isOpen={isOpen}
+              onClose={onClose}
+              hexAddress={module?.address}
+              handleModuleSelect={handleModuleSelect}
+            />
+          )}
         </Flex>
-        <Grid
-          gap={8}
-          templateColumns="minmax(300px, 20%) 1fr"
-          overflow="hidden"
-        >
-          {/* Left side */}
-          <FunctionSelectPanel
-            module={module}
-            tab={selectedType}
-            setTab={setSelectedType}
-            selectedFn={selectedFn}
-            setSelectedFn={handleFunctionSelect}
-          />
-          {/* Right side */}
-          <FunctionSelectBody
+        {isMobile ? (
+          <InteractionBodySectionMobile
             module={module}
             selectedFn={selectedFn}
             openDrawer={onOpen}
           />
-        </Grid>
+        ) : (
+          <InteractionBodySection
+            module={module}
+            selectedType={selectedType}
+            setSelectedType={setSelectedType}
+            selectedFn={selectedFn}
+            handleFunctionSelect={handleFunctionSelect}
+            onOpen={onOpen}
+          />
+        )}
       </PageContainer>
       <Box px={{ base: "16px", md: "48px" }}>
         <ModuleSourceCode sourceCode={verificationData?.source} />