+++ src/lib/pages/nft-collections/index.tsx
@@ -1,6 +1,7 @@
 import { useNftConfig, useTierConfig } from "lib/app-provider";
 import PageContainer from "lib/components/PageContainer";
 import { PageHeader } from "lib/components/PageHeader";
+import { CelatoneSeo } from "lib/components/Seo";
 
 import { Collections } from "./components/Collections";
 
@@ -9,6 +10,7 @@ const NftCollections = () => {
   useNftConfig({ shouldRedirect: true });
   return (
     <PageContainer>
+      <CelatoneSeo pageName="NFT Collections" />
       <PageHeader
         title="NFT Collections"
         subtitle="These are the most recently NFT collections created on this network"